requirements.txt
pip>=23.1.2
setuptools>=67.7.2
wheel>=0.40.0

numpy==1.24.4
pandas==1.5.3
matplotlib==3.7.1
absl-py==1.4.0

torch==2.0.1  # stable release, might work better than 2.1 on 3.12
tensorflow==2.13.0  
gymnasium==0.29.1
scikit-learn==1.3.2
tqdm==4.66.2
python-dateutil==2.8.2
requests==2.31.0
rich==13.7.1
ipykernel==6.29.4
notebook==7.1.2
